requirements_setup.txt
aiomqtt          == 2.0.0
pyserial-asyncio == 0.6
easyconfig       == 0.3.2
pydantic         == 2.6.3
smllib           == 1.3
aiohttp          == 3.9.3
